Resolve #296 -- Improves Performance
SetGlobalVariable was removed and replace with faster Lua alternatives.
Higher precision(resolution) Netloop and replaced ElapsedMilliseconds
with TotalMilliseconds for more accurate timing.

Refs #296

+++ GameServerLib/Logic/GameObjects/BaseTurret.cs
@@ -77,7 +77,7 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
             }
         }
 
-        public override void update(long diff)
+        public override void update(float diff)
         {
             if (!IsAttacking)
             {
